improved makefile

+++ source/tools/picture.cpp
@@ -52,10 +52,10 @@ void Picture::draw_square ( unsigned char Red, unsigned char Green, unsigned cha
 
 void Picture::draw_square_outline ( unsigned char Red, unsigned char Green, unsigned char Blue, 
                            Real draw_x_min, Real draw_x_max, Real draw_y_min, Real draw_y_max ) {
-  int left = ( (draw_x_min - x_min) / (x_max - x_min) ) * (Real) Width;
-  int right = ( (draw_x_max - x_min) / (x_max - x_min ) ) * (Real) Width;
-  int bottom = ( (draw_y_min - y_min) / (y_max - y_min) ) * (Real) Height;
-  int top = ( (draw_y_max - y_min) / (y_max - y_min ) )* (Real) Height;
+  int left = (int) ( ( (draw_x_min - x_min) / (x_max - x_min) ) * (Real) Width );
+  int right = (int) ( ( (draw_x_max - x_min) / (x_max - x_min ) ) * (Real) Width );
+  int bottom = (int) ( ( (draw_y_min - y_min) / (y_max - y_min) ) * (Real) Height );
+  int top = (int) ( ( (draw_y_max - y_min) / (y_max - y_min ) )* (Real) Height );
   if ( right + 1 < Width ) ++ right;
   if ( top  + 1 < Height ) ++ top;
   for ( int i = left; i < right; ++ i ) {
